requirements.txt
PyQt6
pymupdf
langchain
langchain-openai
langchain-community
faiss-cpu
openai
tiktoken
python-dotenv
yt-dlp